Multiple lists on the same server
Multiple lists can now exist on the same server. To access different lists, the user needs to access "http://<server address>/<list name>". If the given list name does not exist, the server will automatically create a new, blank list. Otherwise, the server will load the list and send it back to the client.

Changes were also made to the client to reduce the number of requests to the server. Instead of sending back a status code, the server now sends back the changed list in its response. This means that the client no longer needs to get the list contents in another separate request.

Additionally, a bug where unnamed categories could be created (but not removed) was fixed. All unnamed categories will have a default name of "null"; creating an unnamed category twice will remove all items from the first unnamed category.